backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
pydantic>=2.9.2
pillow>=10.4.0
numpy>=1.26.4
opencv-python-headless>=4.10.0.84
scikit-learn>=1.5.2
PyMuPDF>=1.24.10
torch>=2.6.0
torchvision>=0.21.0
easyocr>=1.7.1
